+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [2.1.0] - 2023-09-12
+### Added
+- Statemachine.ExitState(), allows for exiting currentState
+- Statemachine.Initialized, check if the statemachine is initialized
+
 ## [2.0.0] - 2023-06-28
 ### Changed
 - Statemachine is now a monobehaviour, allows for plugging states without code